pyproject.toml
[project]
name = "vendorless-core"
version = "0.1.0"
description = ""
authors = [
    { name = "Liam Bindle", email ="liambindle@gmail.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "click",
    "pyyaml",
    "cookiecutter",
    "jinja2",
    "vendorless-postgres @ file:///home/liam/projects/vendorless/v2/modules/postgres"
]

[tool.poetry]

packages = [
    { include = "vendorless/core", from = "src" },
]

include = [
    "vendorless/core/templates",
]

[tool.poetry.scripts]
vl = "vendorless.core.cli:cli"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"



[tool.poetry.dependencies]
vendorless-postgres = {develop = true}
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"